fix(tests): share one global workspace across ops memory test modules
parallel tests in documents.rs and learn.rs each had their own
OnceLock<PathBuf>, so they bound global::init() to different temp dirs.
when both ran concurrently, the second init() rebind silently discarded
the first module's seeded data, causing:

  direct_document_handlers_roundtrip_through_namespace — list found no docs
  memory_learn_all_uses_all_namespaces_when_none_is_requested — 0 namespaces

fix: introduce ensure_shared_memory_client() in ops/mod.rs with a single
static OnceLock<PathBuf>; both modules now call global::init() with the
same path so concurrent calls hit the same-workspace fast-path and never
rebind the global client.
