+++ src/openhuman/memory/ops/documents.rs
@@ -492,63 +492,12 @@ pub async fn memory_recall_memories(
 
 #[cfg(test)]
 mod tests {
-    use std::path::PathBuf;
-    use std::sync::OnceLock;
-
     use serde_json::json;
-    use tempfile::TempDir;
 
     use super::*;
 
-    /// Held for the whole test: pins `OPENHUMAN_WORKSPACE` at a stable,
-    /// never-torn-down workspace under `TEST_ENV_LOCK`. These tests call
-    /// `memory_init` → `current_workspace_dir` → `Config::load_or_init`, which
-    /// reads that process-global env var; without this, a concurrent
-    /// env-mutating test can swap the var and tear down its tempdir mid-call,
-    /// yielding `SQLITE_IOERR` / config atomic-replace `ENOENT`.
-    struct WorkspaceEnvGuard {
-        _lock: std::sync::MutexGuard<'static, ()>,
-        previous: Option<std::ffi::OsString>,
-    }
-
-    impl WorkspaceEnvGuard {
-        fn set(path: &std::path::Path) -> Self {
-            let lock = crate::openhuman::config::TEST_ENV_LOCK
-                .lock()
-                .unwrap_or_else(|e| e.into_inner());
-            let previous = std::env::var_os("OPENHUMAN_WORKSPACE");
-            std::env::set_var("OPENHUMAN_WORKSPACE", path);
-            Self {
-                _lock: lock,
-                previous,
-            }
-        }
-    }
-
-    impl Drop for WorkspaceEnvGuard {
-        fn drop(&mut self) {
-            if let Some(previous) = self.previous.as_ref() {
-                std::env::set_var("OPENHUMAN_WORKSPACE", previous);
-            } else {
-                std::env::remove_var("OPENHUMAN_WORKSPACE");
-            }
-        }
-    }
-
-    #[must_use]
-    fn ensure_memory_client() -> WorkspaceEnvGuard {
-        static WORKSPACE: OnceLock<PathBuf> = OnceLock::new();
-        let workspace = WORKSPACE.get_or_init(|| {
-            let tmp = TempDir::new().expect("tempdir");
-            let path = tmp.path().join("workspace");
-            std::fs::create_dir_all(&path).expect("workspace dir");
-            std::mem::forget(tmp);
-            path
-        });
-        // Pin the env BEFORE init so config load/save targets the stable dir.
-        let env_guard = WorkspaceEnvGuard::set(workspace);
-        let _ = crate::openhuman::memory::global::init(workspace.clone());
-        env_guard
+    fn ensure_memory_client() {
+        crate::openhuman::memory::ops::ensure_shared_memory_client();
     }
 
     fn unique_namespace(prefix: &str) -> String {
